Crate rusoto_application_autoscaling

With Application Auto Scaling, you can configure automatic scaling for your scalable resources. You can use Application Auto Scaling to accomplish the following tasks:

Define scaling policies to automatically scale your AWS or custom resources

Scale your resources in response to CloudWatch alarms

Schedule one-time or recurring scaling actions

View the history of your scaling events

Application Auto Scaling can scale the following resources:

Amazon ECS services. For more information, see Service Auto Scaling in the Amazon Elastic Container Service Developer Guide.

Amazon EC2 Spot fleets. For more information, see Automatic Scaling for Spot Fleet in the Amazon EC2 User Guide.

Amazon EMR clusters. For more information, see Using Automatic Scaling in Amazon EMR in the Amazon EMR Management Guide.

AppStream 2.0 fleets. For more information, see Fleet Auto Scaling for Amazon AppStream 2.0 in the Amazon AppStream 2.0 Developer Guide.

Provisioned read and write capacity for Amazon DynamoDB tables and global secondary indexes. For more information, see Managing Throughput Capacity Automatically with DynamoDB Auto Scaling in the Amazon DynamoDB Developer Guide.

Amazon Aurora Replicas. For more information, see Using Amazon Aurora Auto Scaling with Aurora Replicas.

Amazon SageMaker endpoint variants. For more information, see Automatically Scaling Amazon SageMaker Models.

Custom resources provided by your own applications or services. More information is available in our GitHub repository.

To learn more about Application Auto Scaling, including information about granting IAM users required permissions for Application Auto Scaling actions, see the Application Auto Scaling User Guide.

If you're using the service, you're probably looking for ApplicationAutoScalingClient and ApplicationAutoScaling.

Structs

Alarm
Represents a CloudWatch alarm associated with a scaling policy.

ApplicationAutoScalingClient
A client for the Application Auto Scaling API.

CustomizedMetricSpecification
Represents a CloudWatch metric of your choosing for a target tracking scaling policy to use with Application Auto Scaling.

To create your customized metric specification:

Add values for each required parameter from CloudWatch. You can use an existing metric, or a new metric that you create. To use your own metric, you must first publish the metric to CloudWatch. For more information, see Publish Custom Metrics in the Amazon CloudWatch User Guide.

Choose a metric that changes proportionally with capacity. The value of the metric should increase or decrease in inverse proportion to the number of capacity units. That is, the value of the metric should decrease when capacity increases.

For more information about CloudWatch, see Amazon CloudWatch Concepts.

DeleteScalingPolicyRequest

DeleteScalingPolicyResponse

DeleteScheduledActionRequest

DeleteScheduledActionResponse

DeregisterScalableTargetRequest

DeregisterScalableTargetResponse

DescribeScalableTargetsRequest

DescribeScalableTargetsResponse

DescribeScalingActivitiesRequest

DescribeScalingActivitiesResponse

DescribeScalingPoliciesRequest

DescribeScalingPoliciesResponse

DescribeScheduledActionsRequest

DescribeScheduledActionsResponse

MetricDimension
Describes the dimension names and values associated with a metric.

PredefinedMetricSpecification
Represents a predefined metric for a target tracking scaling policy to use with Application Auto Scaling.

PutScalingPolicyRequest

PutScalingPolicyResponse

PutScheduledActionRequest

PutScheduledActionResponse

RegisterScalableTargetRequest

RegisterScalableTargetResponse

ScalableTarget
Represents a scalable target.

ScalableTargetAction
Represents the minimum and maximum capacity for a scheduled action.

ScalingActivity
Represents a scaling activity.

ScalingPolicy
Represents a scaling policy to use with Application Auto Scaling.

ScheduledAction
Represents a scheduled action.

StepAdjustment

Represents a step adjustment for a StepScalingPolicyConfiguration. Describes an adjustment based on the difference between the value of the aggregated CloudWatch metric and the breach threshold that you've defined for the alarm.

For the following examples, suppose that you have an alarm with a breach threshold of 50:

To trigger the adjustment when the metric is greater than or equal to 50 and less than 60, specify a lower bound of 0 and an upper bound of 10.

To trigger the adjustment when the metric is greater than 40 and less than or equal to 50, specify a lower bound of -10 and an upper bound of 0.

There are a few rules for the step adjustments for your step policy:

The ranges of your step adjustments can't overlap or have a gap.

At most one step adjustment can have a null lower bound. If one step adjustment has a negative lower bound, then there must be a step adjustment with a null lower bound.

At most one step adjustment can have a null upper bound. If one step adjustment has a positive upper bound, then there must be a step adjustment with a null upper bound.

The upper and lower bound can't be null in the same step adjustment.

StepScalingPolicyConfiguration
Represents a step scaling policy configuration to use with Application Auto Scaling.

TargetTrackingScalingPolicyConfiguration
Represents a target tracking scaling policy configuration to use with Application Auto Scaling.

Enums

DeleteScalingPolicyError
Errors returned by DeleteScalingPolicy

DeleteScheduledActionError
Errors returned by DeleteScheduledAction

DeregisterScalableTargetError
Errors returned by DeregisterScalableTarget

DescribeScalableTargetsError
Errors returned by DescribeScalableTargets

DescribeScalingActivitiesError
Errors returned by DescribeScalingActivities

DescribeScalingPoliciesError
Errors returned by DescribeScalingPolicies

DescribeScheduledActionsError
Errors returned by DescribeScheduledActions

PutScalingPolicyError
Errors returned by PutScalingPolicy

PutScheduledActionError
Errors returned by PutScheduledAction

RegisterScalableTargetError
Errors returned by RegisterScalableTarget

Traits

ApplicationAutoScaling
Trait representing the capabilities of the Application Auto Scaling API. Application Auto Scaling clients implement this trait.