Struct rusoto_application_autoscaling::PutScheduledActionRequest

pub struct PutScheduledActionRequest {
    pub end_time: Option<f64>,
    pub resource_id: String,
    pub scalable_dimension: Option<String>,
    pub scalable_target_action: Option<ScalableTargetAction>,
    pub schedule: Option<String>,
    pub scheduled_action_name: String,
    pub service_namespace: String,
    pub start_time: Option<f64>,
}

Fields
end_time: Option<f64>

The date and time for the scheduled action to end.

resource_id: String

The identifier of the resource associated with the scheduled action. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

scalable_dimension: Option<String>

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition.

scalable_target_action: Option<ScalableTargetAction>

The new minimum and maximum capacity. You can set both values or just one. During the scheduled time, if the current capacity is below the minimum capacity, Application Auto Scaling scales out to the minimum capacity. If the current capacity is above the maximum capacity, Application Auto Scaling scales in to the maximum capacity.

schedule: Option<String>

The schedule for this action. The following formats are supported:

At expressions - at(yyyy-mm-ddThh:mm:ss)

Rate expressions - rate(value unit)

Cron expressions - cron(fields)

At expressions are useful for one-time schedules. Specify the time, in UTC.

For rate expressions, value is a positive integer and unit is minute | minutes | hour | hours | day | days.

For more information about cron expressions, see Cron.

scheduled_action_name: String

The name of the scheduled action.

service_namespace: String

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

start_time: Option<f64>

The date and time for the scheduled action to start.