Struct rusoto_application_autoscaling::ScalingActivity

pub struct ScalingActivity {
    pub activity_id: String,
    pub cause: String,
    pub description: String,
    pub details: Option<String>,
    pub end_time: Option<f64>,
    pub resource_id: String,
    pub scalable_dimension: String,
    pub service_namespace: String,
    pub start_time: f64,
    pub status_code: String,
    pub status_message: Option<String>,
}

Represents a scaling activity.

Fields
activity_id: String

The unique identifier of the scaling activity.

cause: String

A simple description of what caused the scaling activity to happen.

description: String

A simple description of what action the scaling activity intends to accomplish.

details: Option<String>

The details about the scaling activity.

end_time: Option<f64>

The Unix timestamp for when the scaling activity ended.

resource_id: String

The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

Amazon SageMaker endpoint variants - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider.

scalable_dimension: String

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an Amazon SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

service_namespace: String

The namespace of the AWS service that provides the resource or custom-resource for a resource provided by your own application or service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

start_time: f64

The Unix timestamp for when the scaling activity began.

status_code: String

Indicates the status of the scaling activity.

status_message: Option<String>

A simple message about the current status of the scaling activity.