Trait rusoto_application_autoscaling::ApplicationAutoScaling

pub trait ApplicationAutoScaling {
    fn delete_scaling_policy(
        &self, 
        input: DeleteScalingPolicyRequest
    ) -> RusotoFuture<DeleteScalingPolicyResponse, DeleteScalingPolicyError>;

    fn delete_scheduled_action(
        &self, 
        input: DeleteScheduledActionRequest
    ) -> RusotoFuture<DeleteScheduledActionResponse, DeleteScheduledActionError>;

    fn deregister_scalable_target(
        &self, 
        input: DeregisterScalableTargetRequest
    ) -> RusotoFuture<DeregisterScalableTargetResponse, DeregisterScalableTargetError>;

    fn describe_scalable_targets(
        &self, 
        input: DescribeScalableTargetsRequest
    ) -> RusotoFuture<DescribeScalableTargetsResponse, DescribeScalableTargetsError>;

    fn describe_scaling_activities(
        &self, 
        input: DescribeScalingActivitiesRequest
    ) -> RusotoFuture<DescribeScalingActivitiesResponse, DescribeScalingActivitiesError>;

    fn describe_scaling_policies(
        &self, 
        input: DescribeScalingPoliciesRequest
    ) -> RusotoFuture<DescribeScalingPoliciesResponse, DescribeScalingPoliciesError>;

    fn describe_scheduled_actions(
        &self, 
        input: DescribeScheduledActionsRequest
    ) -> RusotoFuture<DescribeScheduledActionsResponse, DescribeScheduledActionsError>;

    fn put_scaling_policy(
        &self, 
        input: PutScalingPolicyRequest
    ) -> RusotoFuture<PutScalingPolicyResponse, PutScalingPolicyError>;

    fn put_scheduled_action(
        &self, 
        input: PutScheduledActionRequest
    ) -> RusotoFuture<PutScheduledActionResponse, PutScheduledActionError>;

    fn register_scalable_target(
        &self, 
        input: RegisterScalableTargetRequest
    ) -> RusotoFuture<RegisterScalableTargetResponse, RegisterScalableTargetError>;
}

Trait representing the capabilities of the Application Auto Scaling API. Application Auto Scaling clients implement this trait.

Required methods

fn delete_scaling_policy(
&self,
input: DeleteScalingPolicyRequest
) -> RusotoFuture<DeleteScalingPolicyResponse, DeleteScalingPolicyError>

Deletes the specified Application Auto Scaling scaling policy.

Deleting a policy deletes the underlying alarm action, but does not delete the CloudWatch alarm associated with the scaling policy, even if it no longer has an associated action.

To create a scaling policy or update an existing one, see PutScalingPolicy.

fn delete_scheduled_action(
&self,
input: DeleteScheduledActionRequest
) -> RusotoFuture<DeleteScheduledActionResponse, DeleteScheduledActionError>

Deletes the specified Application Auto Scaling scheduled action.

fn deregister_scalable_target(
&self,
input: DeregisterScalableTargetRequest
) -> RusotoFuture<DeregisterScalableTargetResponse, DeregisterScalableTargetError>

Deregisters a scalable target.

Deregistering a scalable target deletes the scaling policies that are associated with it.

To create a scalable target or update an existing one, see RegisterScalableTarget.

fn describe_scalable_targets(
&self,
input: DescribeScalableTargetsRequest
) -> RusotoFuture<DescribeScalableTargetsResponse, DescribeScalableTargetsError>

Gets information about the scalable targets in the specified namespace.

You can filter the results using the ResourceIds and ScalableDimension parameters.

To create a scalable target or update an existing one, see RegisterScalableTarget. If you are no longer using a scalable target, you can deregister it using DeregisterScalableTarget.

fn describe_scaling_activities(
&self,
input: DescribeScalingActivitiesRequest
) -> RusotoFuture<DescribeScalingActivitiesResponse, DescribeScalingActivitiesError>

Provides descriptive information about the scaling activities in the specified namespace from the previous six weeks.

You can filter the results using the ResourceId and ScalableDimension parameters.

Scaling activities are triggered by CloudWatch alarms that are associated with scaling policies. To view the scaling policies for a service namespace, see DescribeScalingPolicies. To create a scaling policy or update an existing one, see PutScalingPolicy.

fn describe_scaling_policies(
&self,
input: DescribeScalingPoliciesRequest
) -> RusotoFuture<DescribeScalingPoliciesResponse, DescribeScalingPoliciesError>

Describes the scaling policies for the specified service namespace.

You can filter the results using the ResourceId, ScalableDimension, and PolicyNames parameters.

To create a scaling policy or update an existing one, see PutScalingPolicy. If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy.

fn describe_scheduled_actions(
&self,
input: DescribeScheduledActionsRequest
) -> RusotoFuture<DescribeScheduledActionsResponse, DescribeScheduledActionsError>

Describes the scheduled actions for the specified service namespace.

You can filter the results using the ResourceId, ScalableDimension, and ScheduledActionNames parameters.

To create a scheduled action or update an existing one, see PutScheduledAction. If you are no longer using a scheduled action, you can delete it using DeleteScheduledAction.

fn put_scaling_policy(
&self,
input: PutScalingPolicyRequest
) -> RusotoFuture<PutScalingPolicyResponse, PutScalingPolicyError>

Creates or updates a policy for an Application Auto Scaling scalable target.

Each scalable target is identified by a service namespace, resource ID, and scalable dimension. A scaling policy applies to the scalable target identified by those three attributes. You cannot create a scaling policy until you have registered the resource as a scalable target using RegisterScalableTarget.

To update a policy, specify its policy name and the parameters that you want to change. Any parameters that you don't specify are not changed by this update request.

You can view the scaling policies for a service namespace using DescribeScalingPolicies. If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy.

Multiple scaling policies can be in force at the same time for the same scalable target. You can have one or more target tracking scaling policies, one or more step scaling policies, or both. However, there is a chance that multiple policies could conflict, instructing the scalable target to scale out or in at the same time. Application Auto Scaling gives precedence to the policy that provides the largest capacity for both scale in and scale out. For example, if one policy increases capacity by 3, another policy increases capacity by 200 percent, and the current capacity is 10, Application Auto Scaling uses the policy with the highest calculated capacity (200% of 10 = 20) and scales out to 30.

Learn more about how to work with scaling policies in the Application Auto Scaling User Guide.

fn put_scheduled_action(
&self,
input: PutScheduledActionRequest
) -> RusotoFuture<PutScheduledActionResponse, PutScheduledActionError>

Creates or updates a scheduled action for an Application Auto Scaling scalable target.

Each scalable target is identified by a service namespace, resource ID, and scalable dimension. A scheduled action applies to the scalable target identified by those three attributes. You cannot create a scheduled action until you have registered the resource as a scalable target using RegisterScalableTarget.

To update an action, specify its name and the parameters that you want to change. If you don't specify start and end times, the old values are deleted. Any other parameters that you don't specify are not changed by this update request.

You can view the scheduled actions using DescribeScheduledActions. If you are no longer using a scheduled action, you can delete it using DeleteScheduledAction.

Learn more about how to work with scheduled actions in the Application Auto Scaling User Guide.

fn register_scalable_target(
&self,
input: RegisterScalableTargetRequest
) -> RusotoFuture<RegisterScalableTargetResponse, RegisterScalableTargetError>

Registers or updates a scalable target. A scalable target is a resource that Application Auto Scaling can scale out and scale in. Each scalable target has a resource ID, scalable dimension, and namespace, as well as values for minimum and maximum capacity.

After you register a scalable target, you do not need to register it again to use other Application Auto Scaling operations. To see which resources have been registered, use DescribeScalableTargets. You can also view the scaling policies for a service namespace using DescribeScalableTargets.

If you no longer need a scalable target, you can deregister it using DeregisterScalableTarget.

Loading content...