Struct rusoto_application_autoscaling::ApplicationAutoScalingClient

pub struct ApplicationAutoScalingClient { /* fields omitted */ }

A client for the Application Auto Scaling API.

Methods

impl ApplicationAutoScalingClient

pub fn new(region: Region) -> ApplicationAutoScalingClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ApplicationAutoScalingClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,